+++ app/controllers/api/ui/base_controller.rb
@@ -35,10 +35,6 @@ class Api::Ui::BaseController < ApplicationController
     puts "#{log_level}: #{message}" if Rails.env.development?
   end
 
-  # def json_request
-  #   request.format.json?
-  # end
-
   def simulate_delay_for_development
     sleep(rand(0.0..1.5)) if Rails.env.development?
   end
@@ -46,16 +42,5 @@ class Api::Ui::BaseController < ApplicationController
   def localized_attr_name
     @localized_attr_name ||= I18n.locale.to_s.downcase == 'it' ? :name_it : :name
   end
-
-  #
-  # Metodi di Papertrail per salvare le modifiche e scrivere il customer loggato che le ha fatte
-  #
-  # def info_for_paper_trail
-  #   {whodunnit_type: 'Customer'}
-  # end
-  #
-  # def user_for_paper_trail
-  #   current_customer.id if current_customer
-  # end
 
 end