+++ app/controllers/application_controller.rb
@@ -1,10 +1,12 @@
 class ApplicationController < ActionController::Base
-  protect_from_forgery with: :exception, unless: :json_request
+  protect_from_forgery with: :exception, unless: :json_request?
   before_action :set_locale
+
+  respond_to :html, :json
 
   protected
 
-  def json_request
+  def json_request?
     request.format.json?
   end