+++ db/migrate/20200111165824_create_users.rb
@@ -3,7 +3,7 @@ class CreateUsers < ActiveRecord::Migration[6.0]
     create_table :users do |t|
       t.string :firstname, limit: 30
       t.string :lastname, limit: 30
-      t.date :birtdate
+      t.date :birthdate
       t.string :gender, limit: 1
       t.integer :state, limit: 1, default: 10
       t.string :phone, limit: 15